+++ back/controllers/comment.js
@@ -30,27 +30,31 @@ exports.ViewAllComment = (req, res) => {
 // Créer un commentaire
 module.exports.createComment = async (req, res) => {
   try {
-    const comment = await Comment.create({
-      id_article: req.body.id_article,
-      id_user: req.auth.userId,
-      titre: req.body.titre,
-      text: req.body.text,
-    });
-    res.status(201).json("Comment created");
-  } catch (error) {
-    console.log(error);
-  }
-};
+  await Comment.create({
+    id_article: req.body.id_article,
+    id_user: req.auth.userId,
+    titre: req.body.titre,
+    text: req.body.text,
+  });
+  res.status(201).json("Comment created");
+} catch (error) {
+  console.log(error);
+  res.status(401).json("Comment not created");
+}
+}
 
 // Modifier titre et/ou texte d'un commentaire
 module.exports.updateComment = async (req, res) => {
+    // Verifie que l'id du commentaire existe
   let com = await Comment.findOne({
     where: { id_commentaire: req.params.id },
   });
   if (!com) {
     return res.status(404).json({ err: "Comment undefined" });
   }
+  // Acces admin ou utilisateur qui a créer le post
   if (req.auth.userId == com.id_user || req.auth.role == 1) {
+    // Mise à jour du commentaire
     Comment.update(
       { titre: req.body.titre, text: req.body.text, date_mod: new Date() },
       {
@@ -65,10 +69,23 @@ module.exports.updateComment = async (req, res) => {
 
 // Suprimer un commentaire
 exports.deleteComment = async (req, res) => {
-  await Comment.destroy({
-    where: {
-      id_commentaire: req.params.id,
-    },
+  // Verifie que l'id du commentaire existe
+  let com = await Comment.findOne({
+    where: { id_commentaire: req.params.id },
   });
-  res.status(201).json("Comment deleted with success");
+  if (!com) {
+    return res.status(404).json({ err: "Comment undefined" });
+  }
+  // Acces admin ou utilisateur qui a créer le post
+  if (req.auth.userId == com.id_user || req.auth.role == 1) {
+    // Supression
+    await Comment.destroy({
+      where: {
+        id_commentaire: req.params.id,
+      },
+    });
+    res.status(201).json("Comment deleted with success");
+  } else {
+    res.status(401).json({ error: "Request non authorized" });
+  }
 };
